v4/services/oracle/oracle_service/server.py
"""
NPS V4 Oracle Service — gRPC server.

Wraps V3 engines (fc60, numerology, oracle, timing_advisor) with gRPC interface.
These engines are near-zero rewrite — they are pure computation with no I/O.

To run:
    python -m oracle_service.server
"""

import logging
import time
from concurrent import futures

# TODO: Import generated protobuf stubs
# from oracle_service.grpc_gen import oracle_pb2, oracle_pb2_grpc

logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s [%(levelname)s] %(name)s: %(message)s",
)
logger = logging.getLogger(__name__)

# Service start time for uptime tracking
_start_time = time.time()


class OracleServiceImpl:
    """Implements oracle.proto OracleService.

    Each method wraps a V3 engine function:
    - GetReading -> oracle.read_sign()
    - GetNameReading -> oracle.read_name()
    - GetQuestionSign -> oracle.question_sign()
    - GetDailyInsight -> oracle.daily_insight()
    - SuggestRange -> timing_advisor + range_optimizer
    - AnalyzeSession -> learner.learn()
    - GetTimingAlignment -> timing_advisor.get_current_quality()
    - HealthCheck -> returns service status
    """

    def GetReading(self, request, context):
        """Get a full oracle reading."""
        # TODO: Import and call engines.oracle.read_sign()
        logger.info("GetReading called")
        raise NotImplementedError

    def GetNameReading(self, request, context):
        """Get a name cipher reading."""
        # TODO: Import and call engines.oracle.read_name()
        logger.info("GetNameReading called")
        raise NotImplementedError

    def GetQuestionSign(self, request, context):
        """Get a question sign."""
        # TODO: Import and call engines.oracle.question_sign()
        logger.info("GetQuestionSign called")
        raise NotImplementedError

    def GetDailyInsight(self, request, context):
        """Get daily insight."""
        # TODO: Import and call engines.oracle.daily_insight()
        logger.info("GetDailyInsight called")
        raise NotImplementedError

    def SuggestRange(self, request, context):
        """Suggest optimal scan range."""
        # TODO: Import and call logic.timing_advisor + logic.range_optimizer
        logger.info("SuggestRange called")
        raise NotImplementedError

    def AnalyzeSession(self, request, context):
        """Analyze a completed session with AI."""
        # TODO: Import and call engines.learner.learn()
        logger.info("AnalyzeSession called")
        raise NotImplementedError

    def GetTimingAlignment(self, request, context):
        """Get cosmic timing alignment."""
        # TODO: Import and call logic.timing_advisor.get_current_quality()
        logger.info("GetTimingAlignment called")
        raise NotImplementedError

    def HealthCheck(self, request, context):
        """Service health check."""
        uptime = int(time.time() - _start_time)
        # TODO: Return HealthResponse protobuf
        logger.info(f"HealthCheck: uptime={uptime}s")
        raise NotImplementedError


def serve(port: int = 50052):
    """Start the gRPC server."""
    import grpc

    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))

    # TODO: Add service to server
    # oracle_pb2_grpc.add_OracleServiceServicer_to_server(OracleServiceImpl(), server)

    # TODO: Add health checking service
    # from grpc_health.v1 import health_pb2_grpc, health
    # health_servicer = health.HealthServicer()
    # health_pb2_grpc.add_HealthServicer_to_server(health_servicer, server)

    server.add_insecure_port(f"[::]:{port}")
    server.start()
    logger.info(f"Oracle service listening on port {port}")
    server.wait_for_termination()


if __name__ == "__main__":
    serve()
